% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindAcquisitionPlans.R
\name{FindAcquisitionPlans}
\alias{FindAcquisitionPlans}
\title{FUNCTION_TITLE}
\usage{
FindAcquisitionPlans(aoi, currentAcquisitionPlans, all = FALSE)
}
\arguments{
\item{aoi}{PARAM_DESCRIPTION}

\item{currentAcquisitionPlans}{PARAM_DESCRIPTION}

\item{all}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_zm}},\code{\link[sf]{st_crs}},\code{\link[sf]{st_transform}},\code{\link[sf]{geos_binary_pred}},\code{\link[sf]{st_coordinates}},\code{\link[sf]{geos_unary}}
\code{\link[lutz]{tz_lookup_coords}}
}
