% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRelativeOrbits.R
\name{FindRelativeOrbits}
\alias{FindRelativeOrbits}
\title{FUNCTION_TITLE}
\source{
\url{http://somewhere.important.com/}
}
\usage{
FindRelativeOrbits(aoi, orbits = S2_relative_orbits)
}
\arguments{
\item{aoi}{PARAM_DESCRIPTION}

\item{orbits}{PARAM_DESCRIPTION, Default: S2_relative_orbits}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_crs}},\code{\link[sf]{st_transform}},\code{\link[sf]{s2}},\code{\link[sf]{geos_binary_pred}},\code{\link[sf]{st_coordinates}},\code{\link[sf]{geos_unary}},\code{\link[sf]{st_geometry}}
\code{\link[lutz]{tz_lookup_coords}}
}
