% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastOverpasses.R
\name{ForecastOverpasses}
\alias{ForecastOverpasses}
\title{FUNCTION_TITLE}
\source{
\url{http://somewhere.important.com/}
}
\usage{
ForecastOverpasses(aoi, period)
}
\arguments{
\item{aoi}{PARAM_DESCRIPTION}

\item{period}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
