% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{GetBootomRight}
\alias{GetBootomRight}
\title{FUNCTION_TITLE}
\source{
\url{http://somewhere.important.com/}
}
\usage{
GetBootomRight(p)
}
\arguments{
\item{p}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_coordinates}}
}
