% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseKML.R
\name{ParseKML}
\alias{ParseKML}
\title{FUNCTION_TITLE}
\usage{
ParseKML(kml)
}
\arguments{
\item{kml}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[xml2]{read_xml}},\code{\link[xml2]{xml_ns_strip}},\code{\link[xml2]{xml_find_all}},\code{\link[xml2]{xml_children}},\code{\link[xml2]{xml_text}},\code{\link[xml2]{as_list}}
\code{\link[sf]{st_as_sf}}
}
