% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTiles.R
\name{PlotTiles}
\alias{PlotTiles}
\title{FUNCTION_TITLE}
\source{
\url{http://somewhere.important.com/}
}
\usage{
PlotTiles(
  x,
  aoi,
  tile_col = NA,
  tile_border = "blue",
  aoi_col = "red",
  aoi_border = aoi_col,
  labels_col = "orange",
  labels_cex = 0.5,
  world = FALSE,
  ...
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{aoi}{PARAM_DESCRIPTION}

\item{tile_col}{PARAM_DESCRIPTION, Default: NA}

\item{tile_border}{PARAM_DESCRIPTION, Default: 'blue'}

\item{aoi_col}{PARAM_DESCRIPTION, Default: 'red'}

\item{aoi_border}{PARAM_DESCRIPTION, Default: aoi_col}

\item{labels_col}{PARAM_DESCRIPTION, Default: 'orange'}

\item{labels_cex}{PARAM_DESCRIPTION, Default: 0.5}

\item{world}{PARAM_DESCRIPTION, Default: FALSE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_geometry}}
\code{\link[maps]{map}}
}
