% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateCurrentAcquisitionPlans.R
\name{UpdateCurrentAcquisitionPlans}
\alias{UpdateCurrentAcquisitionPlans}
\title{FUNCTION_TITLE}
\usage{
UpdateCurrentAcquisitionPlans(past = FALSE, verbose = FALSE, keep = TRUE)
}
\arguments{
\item{past}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{keep}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
